refactor(04-02): move make_tx_pattern/verify_payload_words to test_common.hpp
- Added make_tx_pattern and verify_payload_words inline helpers to test_common.hpp
- Added #include <vector> and #include <gtest/gtest.h> to test_common.hpp
- Removed duplicate anonymous namespace definitions from unicast_runner.cpp
- Removed duplicate anonymous namespace definitions from multicast_runner.cpp
- Runners now resolve both helpers via tt::tt_fabric::test namespace from shared header
